backend/requirements.txt
fastapi[standard]
mysql-connector-python
passlib[bcrypt]
argon2_cffi